+++ CHANGELOG.md
@@ -5,7 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.2] - Unreleased
+
+### Added
+
+- JSON Export Button on Actor Custom Animations (#3)
+- `opacity`, `size`, `loopProperty`, and `animateProperty` effect options.
+
+  #### Animations
+  - Rage
+
+### Fixed
+
+- Multiple effects chained together inheriting previous animations options, for some reason.
 
 ## [0.2.1] - 2024-08-07